requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
pandas==2.2.2
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-experimental==0.3.0
streamlit==1.39.0
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
certifi==2024.8.30
